+++ src/messenger/utilities.ts
@@ -1,5 +1,5 @@
-import type { MessengerTransferData } from '@typing/messenger.js';
+import type { MessengerRawData } from '@typing/messenger.js';
 
-export const isMessengerTransferObject = (data: MessengerTransferData): data is MessengerTransferData => {
+export const isRawMessengerObject = (data: MessengerRawData): data is MessengerRawData => {
     return !!data?.__messengerID && typeof data?.__messengerID === 'string';
 };